fix: address clippy warnings and install fontconfig for CI docs
- Remove unnecessary u32 casts in text.rs
- Use io::Error::other() instead of io::Error::new(io::ErrorKind::Other, ...)
- Allow too_many_arguments for add_quad
- Remove explicit deref and clone on Copy type
- Use Range::contains and is_multiple_of
- Install libfontconfig1-dev for doc build on Ubuntu

+++ crates/motif_core/src/text.rs
@@ -159,7 +159,7 @@ impl TextLayout {
                         let glyphs: Vec<PositionedGlyph> = run
                             .positioned_glyphs()
                             .map(|g| PositionedGlyph {
-                                id: g.id as u32,
+                                id: g.id,
                                 x: g.x,
                                 y: g.y,
                                 advance: g.advance,
@@ -339,7 +339,7 @@ impl TextLayout {
                         let glyphs: Vec<PositionedGlyph> = run
                             .positioned_glyphs()
                             .map(|g| PositionedGlyph {
-                                id: g.id as u32,
+                                id: g.id,
                                 x: g.x,
                                 y: g.y,
                                 advance: g.advance,
